packages/obp-api/obp_api-0.1.0a1.tar.gz/obp_api-0.1.0a1/src/obp_api/resources/products/cascade.py
# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.

from __future__ import annotations

import httpx

from ..._types import NOT_GIVEN, Body, Query, Headers, NoneType, NotGiven
from ..._compat import cached_property
from ..._resource import SyncAPIResource, AsyncAPIResource
from ..._response import (
    to_raw_response_wrapper,
    to_streamed_response_wrapper,
    async_to_raw_response_wrapper,
    async_to_streamed_response_wrapper,
)
from ..._base_client import make_request_options

__all__ = ["CascadeResource", "AsyncCascadeResource"]


class CascadeResource(SyncAPIResource):
    @cached_property
    def with_raw_response(self) -> CascadeResourceWithRawResponse:
        """
        This property can be used as a prefix for any HTTP method call to return the
        the raw response object instead of the parsed content.

        For more information, see https://www.github.com/nemozak1/obp-api-python#accessing-raw-response-data-eg-headers
        """
        return CascadeResourceWithRawResponse(self)

    @cached_property
    def with_streaming_response(self) -> CascadeResourceWithStreamingResponse:
        """
        An alternative to `.with_raw_response` that doesn't eagerly read the response body.

        For more information, see https://www.github.com/nemozak1/obp-api-python#with_streaming_response
        """
        return CascadeResourceWithStreamingResponse(self)

    def delete(
        self,
        product_code: str,
        *,
        bank_id: str,
        # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
        # The extra values given here take precedence over values defined on the client or passed to this method.
        extra_headers: Headers | None = None,
        extra_query: Query | None = None,
        extra_body: Body | None = None,
        timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
    ) -> None:
        """
        <p>Delete a Product Cascade specified by PRODUCT_CODE.</p><p>Authentication is Mandatory</p>

        Args:
          extra_headers: Send extra headers

          extra_query: Add additional query parameters to the request

          extra_body: Add additional JSON properties to the request

          timeout: Override the client-level default timeout for this request, in seconds
        """
        if not bank_id:
            raise ValueError(f"Expected a non-empty value for `bank_id` but received {bank_id!r}")
        if not product_code:
            raise ValueError(f"Expected a non-empty value for `product_code` but received {product_code!r}")
        extra_headers = {"Accept": "*/*", **(extra_headers or {})}
        return self._delete(
            f"/obp/v5.1.0/management/cascading/banks/{bank_id}/products/{product_code}",
            options=make_request_options(
                extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
            ),
            cast_to=NoneType,
        )


class AsyncCascadeResource(AsyncAPIResource):
    @cached_property
    def with_raw_response(self) -> AsyncCascadeResourceWithRawResponse:
        """
        This property can be used as a prefix for any HTTP method call to return the
        the raw response object instead of the parsed content.

        For more information, see https://www.github.com/nemozak1/obp-api-python#accessing-raw-response-data-eg-headers
        """
        return AsyncCascadeResourceWithRawResponse(self)

    @cached_property
    def with_streaming_response(self) -> AsyncCascadeResourceWithStreamingResponse:
        """
        An alternative to `.with_raw_response` that doesn't eagerly read the response body.

        For more information, see https://www.github.com/nemozak1/obp-api-python#with_streaming_response
        """
        return AsyncCascadeResourceWithStreamingResponse(self)

    async def delete(
        self,
        product_code: str,
        *,
        bank_id: str,
        # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
        # The extra values given here take precedence over values defined on the client or passed to this method.
        extra_headers: Headers | None = None,
        extra_query: Query | None = None,
        extra_body: Body | None = None,
        timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
    ) -> None:
        """
        <p>Delete a Product Cascade specified by PRODUCT_CODE.</p><p>Authentication is Mandatory</p>

        Args:
          extra_headers: Send extra headers

          extra_query: Add additional query parameters to the request

          extra_body: Add additional JSON properties to the request

          timeout: Override the client-level default timeout for this request, in seconds
        """
        if not bank_id:
            raise ValueError(f"Expected a non-empty value for `bank_id` but received {bank_id!r}")
        if not product_code:
            raise ValueError(f"Expected a non-empty value for `product_code` but received {product_code!r}")
        extra_headers = {"Accept": "*/*", **(extra_headers or {})}
        return await self._delete(
            f"/obp/v5.1.0/management/cascading/banks/{bank_id}/products/{product_code}",
            options=make_request_options(
                extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
            ),
            cast_to=NoneType,
        )


class CascadeResourceWithRawResponse:
    def __init__(self, cascade: CascadeResource) -> None:
        self._cascade = cascade

        self.delete = to_raw_response_wrapper(
            cascade.delete,
        )


class AsyncCascadeResourceWithRawResponse:
    def __init__(self, cascade: AsyncCascadeResource) -> None:
        self._cascade = cascade

        self.delete = async_to_raw_response_wrapper(
            cascade.delete,
        )


class CascadeResourceWithStreamingResponse:
    def __init__(self, cascade: CascadeResource) -> None:
        self._cascade = cascade

        self.delete = to_streamed_response_wrapper(
            cascade.delete,
        )


class AsyncCascadeResourceWithStreamingResponse:
    def __init__(self, cascade: AsyncCascadeResource) -> None:
        self._cascade = cascade

        self.delete = async_to_streamed_response_wrapper(
            cascade.delete,
        )
